requirements.txt
antlr4-python3-runtime==4.9.2
click==8.0.4
Flask==2.0.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
pathlib==1.0.1
python-dotenv==0.19.2
sidekick-agent-python==0.0.3
six==1.16.0
typing_extensions==4.1.1
websocket-client==1.3.1
Werkzeug==2.0.3
zipp==3.7.0
